Add experimental control over the gaze direction
If this works we may want to finese it a bit, engaging a downward tilt
only as the standard gaze direction goes below the horizon, and then
increasing the rate of change as the angle goes further down.

The GazeManager is so wired in with the rest of the HoloToolkit that I've
decided in this case we should modify the sources instead of providing
our own implementation class.